# storage/api-docs/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Swagger - Games-Searcher",
        "description": "Api documentation for Games-Searcher Laravel project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api/"
        }
    ],
...
    "paths": {
...
        "/games/{gameId}": {
            "get": {
                "tags": [
                    "Game"
                ],
                "summary": "Find game by ID",
                "description": "Display the specified game resource",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "gameId",
                        "in": "path",
                        "description": "ID of game to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game was fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Game"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Game not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Game"
                ],
                "summary": "Update an existing game",
                "description": "Update the specified game resource in storage",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "gameId",
                        "in": "path",
                        "description": "ID of game that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "JSON data for update an existent game in database",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Game"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful update",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Game"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Game not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Game"
                ],
                "summary": "Delete game",
                "description": "Remove the specified game resource from storage",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "gameId",
                        "in": "path",
                        "description": "ID of the game that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid game id supplied"
                    },
                    "404": {
                        "description": "Game not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Game": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "10"
                    },
                    "name": {
                        "type": "string",
                        "example": "Minecraft"
                    },
                    "rating": {
                        "type": "number",
                        "example": 10.5
                    },
                    "summary": {
                        "type": "string",
                        "example": "Top game for kids"
                    },
                    "first_release_date": {
                        "type": "string",
                        "example": "2022-01-12"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "required": [
                    "id",
                    "name",
                    "category_id",
                    "rating",
                    "summary",
                    "first_release_date"
                ]
            }
        }
    }
}